pyproject.toml
[tool.poetry]
name = "corrective-rag-langgraph"
version = "0.1.0"
description = "A corrective RAG system built with LangGraph"
authors = ["Robin Ochieng <your.email@example.com>"]
readme = "README.md"
packages = [{include = "corrective_rag", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
beautifulsoup4 = "*"
langchain-community = "*"
langchain-chroma = "*"
tiktoken = "*"
langchainhub = "*"
langchain = "*"
langgraph = "*"
tavily-python = "*"
langchain-openai = "*"
python-dotenv = "*"
faiss-cpu = "*"
pypdf = "*"
langchain-text-splitters = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88